Update capture_timelapse.py
Did a manual "hard reset" by copy-pasting in yesterday's version. I tried using the git terminal commands, but it wasn't having it. I will work on re-integrating today's changes (improvement) when I have time to test them further. I should not have merged the experiment branch as early as I did. It worked, but I was unhappy with it.